+++ buffer.h
@@ -1,24 +1,26 @@
 #ifndef BUFFER_H
 #define BUFFER_H
 
+#include <sys/types.h>
+
 typedef struct buffer {
   char *x;
-  unsigned int p;
-  unsigned int n;
+  off_t p;
+  off_t n;
   int fd;
-  int (*op)();
+  ssize_t (*op)();
 } buffer;
 
 #define BUFFER_INIT(op,fd,buf,len) { (buf), 0, (len), (fd), (op) }
 #define BUFFER_INSIZE 8192
 #define BUFFER_OUTSIZE 8192
 
-extern void buffer_init(buffer *,int (*)(),int,char *,unsigned int);
+extern void buffer_init(buffer *,ssize_t (*)(),int,char *,off_t);
 
 extern int buffer_flush(buffer *);
-extern int buffer_put(buffer *,char *,unsigned int);
-extern int buffer_putalign(buffer *,char *,unsigned int);
-extern int buffer_putflush(buffer *,char *,unsigned int);
+extern int buffer_put(buffer *,char *,off_t);
+extern int buffer_putalign(buffer *,char *,off_t);
+extern int buffer_putflush(buffer *,char *,off_t);
 extern int buffer_puts(buffer *,char *);
 extern int buffer_putsalign(buffer *,char *);
 extern int buffer_putsflush(buffer *,char *);
@@ -29,12 +31,12 @@ extern int buffer_putsflush(buffer *,char *);
     : buffer_put((s),&(c),1) \
   )
 
-extern int buffer_get(buffer *,char *,unsigned int);
-extern int buffer_bget(buffer *,char *,unsigned int);
+extern int buffer_get(buffer *,char *,off_t);
+extern int buffer_bget(buffer *,char *,off_t);
 extern int buffer_feed(buffer *);
 
 extern char *buffer_peek(buffer *);
-extern void buffer_seek(buffer *,unsigned int);
+extern void buffer_seek(buffer *,off_t);
 
 #define buffer_PEEK(s) ( (s)->x + (s)->n )
 #define buffer_SEEK(s,len) ( ( (s)->p -= (len) ) , ( (s)->n += (len) ) )
